fastboot: Updated sparse API with 64-bit length
To support large file, the length data type should be updated
to int64_t from int.

Reviewed-on: http://gerrit.mot.com/687095
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Connie Zhao <czhao1@motorola.com>

+++ fastboot/protocol.cpp
@@ -269,9 +269,9 @@ int64_t fb_upload_data(Transport* transport, const char* outfile) {
 
 #define TRANSPORT_BUF_SIZE 1024
 static char transport_buf[TRANSPORT_BUF_SIZE];
-static int transport_buf_len;
+static int64_t transport_buf_len;
 
-static int fb_download_data_sparse_write(void *priv, const void *data, int len)
+static int fb_download_data_sparse_write(void *priv, const void *data, int64_t len)
 {
     int r;
     Transport* transport = reinterpret_cast<Transport*>(priv);
